Refactor move handling and charting logic
Refactored `SanMoveNumberString` in `Board.fs` to replace redundant methods and improve move number generation. Updated ply count handling for better synchronization with `position.Ply`.

Enhanced move annotations in `Tournament.fs` and `Chessboard.razor` by adopting `SanMoveNumberString`. Improved navigation and comment updates in `Chessboard.razor`.

Refactored chart components (`DualEngineAnalysis.razor`, `SingleEngineAnalysis.razor`) with a new `GamePgnChart` class for evaluation rendering and dynamic updates. Added move indicator support in charts.

Updated `UciOption.fs` to use `int64` for `Spin` options, improving compatibility with larger values. Adjusted parsing and validation logic accordingly.

+++ ChessLibrary/Engine.fs
@@ -561,7 +561,7 @@ module Engine =
               | Some option ->
                   match option.OptionType with        
                   | UciOption.Spin (min, max, _) -> 
-                      let intValue = int ms
+                      let intValue = int64 ms
                       if intValue >= min && intValue <= max then
                         //create a setoption based on the option and the intValue
                         let cmd = sprintf "setoption name %s value %d" option.Name intValue
@@ -895,7 +895,7 @@ module Engine =
         | Some option ->
             match option.OptionType with        
             | UciOption.Spin (min, max, _) -> 
-                let intValue = int milliSeconds
+                let intValue = int64 milliSeconds
                 if intValue >= min && intValue <= max then
                   //create a setoption based on the option and the intValue
                   let cmd = sprintf "setoption name %s value %d" option.Name intValue
